Add support for homemade engines (#34)
Additional changes:
The !name command is printing the engine name along with the version and bot name like it is in lichess-bot. Now it will print "AttackingOrDefending running YaneuraOu (Lishogi-Bot v1.1.0)". The if len(cmds) > 0 in search_with_ponder, is moved to engine_ctrl with "and nodes is None and depth is None and movetime is None". search_for now calls search instead of go. first_search is removed and we are using search_for. print_stats() and get_stats() are changed to be like in lichess-bot (with a few changes). name() is changed to be like in lichess-bot. ponderhit, stop, quit, kill_process are added in EngineWrapper so Homemade engines can inherit them. go_commands is changed to be like in lichess-bot (they weren't working before because it tried to find them inside the 'engines' but it was inside 'engines.usi_options'). is_usi_ponder is renamed to can_ponder and we remove "is_usi and " because homemade engines can ponder. Added the changed for corresponce from lichess-bot to lishogi-bot. Changed play_first_move to choose_first_move which returns the move (the engine will also be able to ponder after the first move). Moved "if is_usi_ponder and ponder_move is not None:" to inside start_pondering like it was after the refactor in lichess-bot (ShailChoksi/lichess-bot#270). Moved engine.print_stats() to search.